Added a few more temp wrappers and rewrote some parts of the old one

+++ LaquaiLib/Classes/TempWrappers/TempFile.cs
@@ -1,6 +1,6 @@
 ﻿using System.IO;
 
-namespace LyricsDisplay.Util;
+namespace LaquaiLib.Util;
 
 /// <summary>
 /// Represents a temporary file that is automatically deleted when its wrapper object is disposed.
@@ -8,13 +8,13 @@ namespace LyricsDisplay.Util;
 public class TempFile : IDisposable
 {
     /// <summary>
-    /// Instantiates a new <see cref="TempFile"/> with a fully random file name and extension as assigned by the OS.
+    /// Instantiates a new <see cref="TempFile"/> with the file name and extension being assigned by the OS. It is usually a <see cref="Guid"/> with the extension <c>.tmp</c>.
     /// </summary>
     public TempFile()
-        : this(System.IO.Path.GetTempFileName()) { }
+        : this(System.IO.Path.GetTempFileName(), null) { }
 
     /// <summary>
-    /// Instantiates a new <see cref="TempFile"/> with a fully random file name as assigned by the OS, and changes its extension to the one specified file extension.
+    /// Instantiates a new <see cref="TempFile"/> with the file name being assigned by the OS (it is usually a <see cref="Guid"/>), and changes its extension to the one specified.
     /// </summary>
     /// <param name="fileExtension">The file extension for this <see cref="TempFile"/>.</param>
     public TempFile(string fileExtension)
@@ -24,15 +24,9 @@ public class TempFile : IDisposable
     /// Instantiates a new <see cref="TempFile"/> as a wrapper around the specified file path. If the target file does not exist, it is created. A deletion attempt is still made when the wrapping <see cref="TempFile"/> is disposed.
     /// </summary>
     /// <param name="path">The path to the file to wrap with this <see cref="TempFile"/>.</param>
-    /// <param name="fileExtension">The file extension for this <see cref="TempFile"/>. This is the extension <paramref name="path"/> is changed to before opening the file stream. If <c>null</c> or white space, the existing extension in <paramref name="path"/> is not altered.</param>
-    /// <exception cref="ObjectDisposedException"></exception>
+    /// <param name="fileExtension">The file extension for this <see cref="TempFile"/>. This is the extension <paramref name="path"/> is changed to before opening the file stream. If <c>null</c> or white space, the existing extension in <paramref name="path"/> is kept.</param>
     public TempFile(string path, string fileExtension)
     {
-        if (string.IsNullOrWhiteSpace(path))
-        {
-            throw new ObjectDisposedException(nameof(_path));
-        }
-
         if (string.IsNullOrWhiteSpace(fileExtension))
         {
             _path = path;
@@ -52,39 +46,43 @@ public class TempFile : IDisposable
     /// </summary>
     public string Path {
         get {
-            if (string.IsNullOrWhiteSpace(_path))
-            {
-                throw new ObjectDisposedException(nameof(Path));
-            }
+            ObjectDisposedException.ThrowIf(string.IsNullOrWhiteSpace(_path) && _stream is null, _path);
             return _path;
         }
     }
-
     /// <summary>
-    /// 
+    /// The <see cref="FileStream"/> for the file this <see cref="TempFile"/> wraps.
     /// </summary>
     public FileStream Stream {
         get {
-            ObjectDisposedException.ThrowIf(_stream is null, _stream);
+            ObjectDisposedException.ThrowIf(string.IsNullOrWhiteSpace(_path) && _stream is null, _stream);
             return _stream;
         }
     }
+    /// <summary>
+    /// Whether this <see cref="TempFile"/> has been disposed.
+    /// </summary>
+    public bool IsDisposed => string.IsNullOrWhiteSpace(_path) && _stream is null;
 
+    #region Dispose pattern
     private void Dispose(bool disposing)
     {
-        if (!string.IsNullOrWhiteSpace(_path))
+        if (disposing)
         {
-            try
+            if (!string.IsNullOrWhiteSpace(_path))
             {
-                File.Delete(_path);
+                try
+                {
+                    File.Delete(_path);
+                }
+                catch { }
+                _path = null;
             }
-            catch { }
-            _path = null;
-        }
-        if (_stream is not null)
-        {
-            _stream.Dispose();
-            _stream = null;
+            if (_stream is not null)
+            {
+                _stream.Dispose();
+                _stream = null;
+            }
         }
     }
 
@@ -99,4 +97,5 @@ public class TempFile : IDisposable
         GC.SuppressFinalize(this);
         Dispose(true);
     }
+    #endregion
 }
